TP5/requirements.txt
matplotlib~=3.4.3
pandas~=1.3.2
numpy~=1.20.1
Pillow~=8.1.2